workspaces/misc/Chamber-of-Echos/pyproject.toml
[project]
name = "chamber-of-echos"
version = "0.1.0"
description = ""
authors = [
    {name = "ShotaTanemura",email = "shota0711tane@gmail.com"}
]
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "pycryptodome (>=3.23.0,<4.0.0)",
    "scapy (>=2.6.1,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
